Move old utilities into separate dir
